backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydub==0.25.1
SpeechRecognition==3.10.4
groq==0.11.0
openai==0.28.0
gTTS==2.5.3
elevenlabs==1.9.0
